fix: Rename file to match component name
